Some initial refactoring of OverlayManager - split WindowContext/Overlaymanager
