+++ browser/src/UI/Overlay/OverlayManager.ts
@@ -1,0 +1,89 @@
+import * as _ from "lodash"
+import { IScreen } from "./../../Screen"
+import { NeovimInstance } from "./../../NeovimInstance"
+
+import { WindowContext } from "./WindowContext"
+
+export interface IOverlay {
+    update(element: HTMLElement, windowContext: WindowContext): void
+}
+
+interface IOverlayInfo {
+    overlay: IOverlay
+    element: HTMLElement
+}
+
+export class OverlayManager {
+
+    private _screen: IScreen
+    private _containerElement: HTMLElement
+
+    private _lastEventContext: Oni.EventContext
+    private _lastWindowData: any
+
+    private _overlays: { [key: string]: IOverlayInfo } = {}
+    private _neovimInstance: NeovimInstance
+
+    constructor(screen: IScreen, neovimInstance: NeovimInstance) {
+        this._screen = screen
+
+        const div = document.createElement("div")
+        div.style.position = "absolute"
+        div.style.top = "0px"
+        div.style.left = "0px"
+        div.style.width = "100px"
+        div.style.height = "100px"
+        document.body.appendChild(div)
+        this._containerElement = div
+
+        this._neovimInstance = neovimInstance
+    }
+
+    public addOverlay(key: string, overlay: IOverlay): void {
+        const overlayContainer = document.createElement("div")
+        overlayContainer.className = "overlay-container"
+
+        this._containerElement.appendChild(overlayContainer)
+
+        this._overlays[key] = {
+            overlay,
+            element: overlayContainer,
+        }
+    }
+
+    public notifyWindowDimensionsChanged(eventContext: Oni.EventContext, data: any) {
+        this._lastEventContext = eventContext
+        this._lastWindowData = data
+
+        this._redrawWithDelay()
+    }
+
+    private _redrawWithDelay(): void {
+        setTimeout(() => this._redrawElement(), 250)
+    }
+
+    private _redrawElement(): void {
+
+        if (!this._lastWindowData || !this._lastEventContext) {
+            return
+        }
+
+        const windowStartRow = this._screen.cursorRow - this._lastEventContext.winline + 1
+        const windowStartColumn = this._screen.cursorColumn - this._lastEventContext.wincol + 1
+
+        this._containerElement.style.top = (windowStartRow * this._screen.fontHeightInPixels) + "px"
+        this._containerElement.style.left = (windowStartColumn * this._screen.fontWidthInPixels) + "px"
+
+        const width = (this._lastWindowData.dimensions.width * this._screen.fontWidthInPixels) + "px"
+        const height = (this._lastWindowData.dimensions.height * this._screen.fontHeightInPixels) + "px"
+
+        this._containerElement.style.width = width
+        this._containerElement.style.height = height
+
+        const windowContext = new WindowContext(this._lastWindowData, this._screen.fontWidthInPixels, this._screen.fontHeightInPixels, this._lastEventContext)
+
+        _.values(this._overlays).forEach((overlayInfo) => {
+            overlayInfo.overlay.update(overlayInfo.element, windowContext)
+        })
+    }
+}
